refactor ticket testing suite
